Refactor and add basic test for stitching
