+++ tests/reflectometry/test1.txt
@@ -1,0 +1,31 @@
+# # ORSO reflectivity data file | 0.1 standard | YAML encoding | https://reflectometry.org
+# creator:
+#   name: scipp
+#   time: '2021-06-30T20:53:53'
+#   system: owens-MBP.home
+# data_source:
+#   experiment_date: '2021-06-30'
+#   experiment:
+#     instrument: AMOR
+#     probe: neutron
+# reduction:
+#   software: ess-v0.0.4-32-gf95d164
+#   comment: Live reduction
+# columns:
+# - quantity: wavelength
+#   unit: "\xC5"
+# - quantity: theta
+#   unit: deg
+# - quantity: Reflectivity
+#   unit: dimensionless
+#   comment: A 2D map with theta in horizontal and wavelength in vertical
+# 
+1.1111111111111111e+01 1.1111111111111111e+01 1.0000000000000000e+00 nan nan nan nan nan nan nan nan
+2.2222222222222221e+01 2.2222222222222221e+01 nan nan nan nan nan nan nan nan nan
+3.3333333333333329e+01 3.3333333333333329e+01 nan nan nan nan nan nan nan nan nan
+4.4444444444444443e+01 4.4444444444444443e+01 nan nan nan nan nan nan nan nan nan
+5.5555555555555557e+01 5.5555555555555557e+01 nan nan nan nan nan nan nan nan nan
+6.6666666666666657e+01 6.6666666666666657e+01 nan nan nan nan nan nan nan nan nan
+7.7777777777777771e+01 7.7777777777777771e+01 nan nan nan nan nan nan nan nan nan
+8.8888888888888886e+01 8.8888888888888886e+01 nan nan nan nan nan nan nan nan nan
+1.0000000000000000e+02 1.0000000000000000e+02 nan nan nan nan nan nan nan nan nan
