+++ tests/reflectometry/test2.txt
@@ -1,0 +1,4 @@
+# hello
+3.6666666666666665e+00 3.3333334326744080e-01 1.9245009124279022e-01 3.2500000000000001e-01
+7.3333333333333330e+00 3.3333334326744080e-01 1.9245009124279022e-01 6.6249999999999998e-01
+1.1000000000000000e+01 3.3333334326744080e-01 1.9245009124279022e-01 1.0000000000000000e+00
